# src/main/java/com/gmail/speedwagon/dto/order/OrderItemResponse.java
package com.gmail.speedwagon.dto.order;

import com.gmail.speedwagon.dto.product.ProductResponse;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
public class OrderItemResponse {
    private Long id;
    private Long amount;
    private Long quantity;
    private ProductResponse product;
}